pyproject.toml
[project]
name = "weichattool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "av>=17.0.0",
    "pandas>=2.3.3",
    "pillow>=12.1.1",
    "pycryptodome>=3.23.0",
    "python-dotenv>=1.2.2",
    "silk-python>=0.2.8",
]
